+++ src/lib.rs
@@ -58,6 +58,8 @@
 //! }
 //! ```
 
+#![cfg_attr(docsrs, feature(doc_cfg))]
+
 use std::ffi::{self, CStr};
 use std::fmt;
 
@@ -88,11 +90,13 @@ pub type TstampType = TimestampType;
 mod raw;
 
 #[cfg(windows)]
+#[cfg_attr(docsrs, doc(cfg(windows)))]
 pub mod sendqueue;
 
 #[cfg(feature = "capture-stream")]
 mod stream;
 #[cfg(feature = "capture-stream")]
+#[cfg_attr(docsrs, doc(cfg(feature = "capture-stream")))]
 pub use stream::PacketStream;
 
 /// An error received from pcap
@@ -227,6 +231,14 @@ impl From<std::io::ErrorKind> for Error {
     }
 }
 
+/// Return size of a commonly used packet header.
+///
+/// On Windows this packet header is implicitly added to send queues, so this size must be known
+/// if an application needs to precalculate the exact send queue buffer size.
+pub const fn packet_header_size() -> usize {
+    std::mem::size_of::<raw::pcap_pkthdr>()
+}
+
 #[cfg(test)]
 mod tests {
     use std::error::Error as StdError;
@@ -278,5 +290,13 @@ mod tests {
                 _ => assert!(error.cause().is_none()),
             }
         }
+    }
+
+    #[test]
+    fn test_packet_size() {
+        assert_eq!(
+            packet_header_size(),
+            std::mem::size_of::<raw::pcap_pkthdr>()
+        );
     }
 }