+++ src/sendqueue/mod.rs
@@ -1,3 +1,5 @@
+//! Utilities for sending batches of packets.
+
 #[cfg(windows)]
 pub mod windows;
 #[cfg(windows)]